Initial and Final commit
Calculated the total donations. Calculated the donations. The total amount of donations is displayed with a thousands separator in the report. Showed only donators that give 1000 dollars or more. Made Major donators in descending order. Created the table for major donators. Created rows for each individual donator. Closed the element tag to finish the table in the donorTable variable. Added The value of the summary table that is displayed in the div element with the id of donorTable. Changed the layout css to look more professional. Ready for review.